+++ tests/back/Pim/Enrichment/EndToEnd/Product/ProductModel/ExternalApi/PartialUpdateListProductModelEndToEnd.php
@@ -268,7 +268,7 @@ JSON;
 
         $expectedContent =
             <<<JSON
-{"line":1,"code":"sub_sweat_option_a","status_code":422,"message":"Validation failed.","errors":[{"property":"attribute","message":"Variant axis \"a_simple_select\" cannot be modified, \"[optionB]\" given"}]}
+{"line":1,"code":"sub_sweat_option_a","status_code":204}
 JSON;
 
         $response = $this->executeStreamRequest('PATCH', 'api/rest/v1/product-models', [], [], [], $data);
@@ -295,24 +295,6 @@ JSON;
         $response = $this->executeStreamRequest('PATCH', 'api/rest/v1/product-models', [], [], [], $data);
         $httpResponse = $response['http_response'];
 
-        $this->assertSame(Response::HTTP_OK, $httpResponse->getStatusCode());
-        $this->assertSame($expectedContent, $response['content']);
-    }
-
-    public function testCreateAndUpdateProductModelsWithUpdatedAxeValue()
-    {
-        $data =
-            <<<JSON
-    {"code": "sub_sweat_option_a", "parent": "sweat", "values": {"a_simple_select": [{"locale": null, "scope": null, "data": "optionB"}]}}
-    {"code": "sub_sweat_option_b", "parent": "sweat", "values": {"a_simple_select": [{"locale": null, "scope": null, "data": "optionB"}]}}
-JSON;
-        $expectedContent =
-            <<<JSON
-{"line":1,"code":"sub_sweat_option_a","status_code":422,"message":"Validation failed.","errors":[{"property":"attribute","message":"Variant axis \"a_simple_select\" cannot be modified, \"[optionB]\" given"}]}
-{"line":2,"code":"sub_sweat_option_b","status_code":201}
-JSON;
-        $response = $this->executeStreamRequest('PATCH', 'api/rest/v1/product-models', [], [], [], $data);
-        $httpResponse = $response['http_response'];
         $this->assertSame(Response::HTTP_OK, $httpResponse->getStatusCode());
         $this->assertSame($expectedContent, $response['content']);
     }